+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.core.BaseballGame;
+
 public class Application {
     public static void main(String[] args) {
         BaseballGame baseballGame = new BaseballGame();
